Aplpy same change to safekeeper stop
